sysctl: rename ->used to ->ctl_use_refs
In a later patch I will split the 'count' counter. We need to have a
clear distinction between the three counters to be able to understand
the code.

This counts the number of references to this object from places that
can tinker with it's internals (e.g. ctl_table, ctl_entry,
attached_to, attached_by, etc.).

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>