sysctl: rename ->used to ->ctl_use_refs
In a later patch I will split the 'count' counter in two. We need to
have a clear distinction between the counters to be able to understand
the code.

This counts the number of references to this object from places that
can tinker with it's internals (e.g. ctl_table, ctl_entry,
attached_to, attached_by, etc.).

The removal of `header->used = 0;` from __register_sysctl_paths_impl
does not change anything, as `header` is allocated with kzalloc.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>