sysctl: add register_sysctl_dir: register an empty sysctl directory
There are a few places in the tree that register a empty ctl_table
array just to make sure the directories are created.

The empty ctl_table takes up memory and is checked at lookup/readdir.

Registering empty dirs with this function only takes care of creating
the directories, without adding the empty ctl_table.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>