sysctl: add ctl_cookie and ctl_cookie_handler
The cookie represent a piece of data and a handler for it that is
associated with the header.

No one uses this at this point. Patches will passing cookie + handler
will be submitted later in various kernel subsystems that can benefit
from this.

The handler is meant to create a new ctl_table based on a template
ctl_table and the cookie. This is useful in lots of places in the
kernel where we kmemdup() an array of template ctl_tables and change
fields in a predictable way.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>