sysctl: always perform sysctl checks
I'm not sure if we should always perform checks without giving
EMBEDDED folks a chance to skip these checks. If the system is
verified with sysctl checks, one can disable them to skip some
performance penalties.

Two of these checks will be more costly than the others:
- sysctl_check_netns_correspondents
- sysctl_check_duplicates

The first check is run for all regular (not netns dependent) sysctl
directories that are registered. For each element of a path
(e.g. kernel/sched_domain/cpu0/domain0/) we scan the current netns's
list of netns correspondents and check for matches.

This is O(len(current-netns-corresp-list)) * O(depth(registered path))

The second check compares every item in a registered table with all
siblings (other tables or subdirectories that are children of the same
parent directory).

This is O(len(table-being-registered)) * O(sum(len(already-registered-tables))+nr-subdirs).

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>