sysctl: register only tables of sysctl files
We're trying to move away from using ctl_tables to encode directories.

The old sysctl registration function can attach a whole sub-tree of
sysctls with directories and tables of files in one single operation.

This patch splits the registration of a complex ctl_table array which
may have arbitrary numbers of directories (->child != NULL) and tables
of files into a series of simpler registrations that only register
tables of files.

Graphically:

   register('dir', { + file-a
                     + file-b
                     + subdir1
                       + file-c
                     + subdir2
                       + file-d
                       + file-e })

is transformed into:
   wrapper->subheaders[0] = register('dir', {file1-a, file1-b})
   wrapper->subheaders[1] = register('dir/subdir1', {file-c})
   wrapper->subheaders[2] = register('dir/subdir2', {file-d, file-e})
   return wrapper

No attempt was made to optimise this function. This hack is only
useful until all in-kernel users that register sysctl tables get rid
of the .child member of ctl_table and manually register arrays of
ctl_table files themselves.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>