sysctl: register only tables of sysctl files
We're trying to move away from using ctl_tables to encode
directories. We split the registration of a complex ctl_table array
which may have directories (.child != NULL) into a series of simpler
registration that only register tables of files.

To maintain full compatibility with users of this function we extend
the ctl_table_header to hold information about the subheaders.

This hack is only useful until all places that register sysctl tables
get rid of the .child member of ctl_table and register only arrays of
ctl_table files themselves.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>