sysctl: reorder members of ctl_table_header (cleanup)
In the next commits entries that are specific to directory
ctl_table_headers and to file-wrapping ctl_table_headers will be put
in unions. Things will get ugly if we keep members in
union-in-struct-in-union-in-...

This makes the structure grow: the .rcu member is kept out of the
union until I understand which fields may still be accessed in
read-side rcu after the header is scheduled for deletion with call_rcu.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>