sysctl: add ctl_type member
We have three types of headers and it's hard to keep track of which is
which and things will get a bit worse later when dirs will be stored
in a rbtree.

This doesn't take up extra space because we had a 2B hole to fill (now
reduced to 1B).

There was a subtle check in unregister_sysctl_table(_impl):
 - dirs and tables are members of their parent's corresponding list
 - netns-specific dirs are members of a netns-specific list of headers

When deleting the header form the list we needed to take the list owner's lock:
 - the parent's lock for dirs/tables
 - netns-specific list protector lock for netns-specific dirs

The decision was (ctl_dirname==NULL) => type = netns-specific dir.

That only made sense because ctl_dirname and ctl_table_arg are unioned
and ctl_table_headers that wrap file arrays never have ctl_table_arg==NULL.

But once the ctl_dirname/ctl_table_arg fields no longer share the same
memory the check will not be valid.

Using ctl_type the check is no longer dependent on that union.

Signed-off-by: Lucian Adrian Grijincu <lucian.grijincu@gmail.com>